refactor(middleware-cors): adjusted middleware to changed Effect interface
